Changed to single-pass bitwise algorithm
Enhanced code to use bitwise operators to track live neighbor counts
for all cells and whether a cell is active or not.  This exponentially
increases performance as it prevents iterating over the entire list for
each cell’s neighbor state.

NOTE: There does seem to be a bug in how browsers handle the bitwise
operation near the edges of the map.  Trying to work that out.